Update search.cam.ac.uk URL to be https:// instead of http://
If one's Drupal site uses https, Chrome (and possibly other browsers) refuse to
mark the page as "secure" because 'The page includes a form with a non-secure
action" attribute' - specifically, the search box. As of Chrome v62 (due for
release October 17th), using any <input> element on a http:// site will lead
to a security warning, hence PL sites may well be aiming to switch to https://
before then.