Fix footer CSS behaviour
Set classes for Study at Cambridge and Research at Cambridge sub sections to match the About the University sub section so that they behave in the same way for mobile (currently the child links for those two just disappear on mobile). This will bring it in line with homepage behaviour and make it more consistent.